+++ changelog.md
@@ -10,6 +10,7 @@ v1.5 (2021-xx-xx):
 * Add new "overrideaiability" command for overriding jump / spin ability AI
   * (advanced, intended for experimentation w/ custom characters - use "bothelp 1" for more info)
 * Fix rare issue where SP bot could reset your rings to 0 when exiting a level
+* Fix issue where SP bot could grant a perfect bonus with only half a level's rings collected (oops)
 * Many minor fixes and code cleanups, eradicating most magic numbers
 
 *AI Behavior Changes:*